Update How the best app sometimes is no app at all.md
This lessons learned entry from Digital Literacy class with Peter Bell contrasts the benefits of a responsive website with the benefits of an app, and within the options to create an app, the benefits of PhoneGap and Titanium with a native app (most expensive option). Finally, it provides a link to the most successful mobile apps and what made them so successful.